refactor(components): split SettingsPanel into per-section subcomponents
Split the 1527-line SettingsPanel.tsx into a thin parent composition
plus 14 per-section subcomponents under components/settings/. Named
and default exports of SettingsPanel are preserved byte-identical; the
import in pages/SettingsPage.tsx continues to work without edits.

Part of #290 (Tier 1, task 11).
