+++ src/crosshook-native/src/components/settings/format.ts
@@ -1,0 +1,45 @@
+/** Returns up to `maxItems` paths from the list, or all paths when `maxItems` is absent. */
+export function toDisplayList(paths: string[], maxItems?: number) {
+  if (!Number.isFinite(maxItems) || !maxItems || maxItems <= 0) {
+    return paths;
+  }
+
+  return paths.slice(0, maxItems);
+}
+
+/** Truncates a path to at most 96 characters, keeping head and tail. */
+export function truncatePath(path: string) {
+  const normalized = path.trim();
+  if (normalized.length <= 96) {
+    return normalized;
+  }
+
+  return `${normalized.slice(0, 40)}...${normalized.slice(-48)}`;
+}
+
+/** Formats a byte count as a human-readable string (e.g. "1.23 GiB"). */
+export function formatBytes(value: number) {
+  if (!Number.isFinite(value) || value <= 0) {
+    return '0 B';
+  }
+  const units = ['B', 'KiB', 'MiB', 'GiB', 'TiB'];
+  let current = value;
+  let unitIndex = 0;
+  while (current >= 1024 && unitIndex < units.length - 1) {
+    current /= 1024;
+    unitIndex += 1;
+  }
+  return `${current.toFixed(current >= 100 ? 0 : current >= 10 ? 1 : 2)} ${units[unitIndex]}`;
+}
+
+/** Formats an ISO timestamp string for display, returning "Unknown" when absent. */
+export function formatTimestamp(value: string | null) {
+  if (!value) {
+    return 'Unknown';
+  }
+  const parsed = new Date(value);
+  if (!Number.isFinite(parsed.getTime())) {
+    return value;
+  }
+  return parsed.toLocaleString();
+}
